+++ client/network/deserializers/lobbyServerDeserializersModule.cpp
@@ -53,7 +53,7 @@ void LobbyServerDeserializersModule::registerDeserializers(
         PlayerDto playerDto{};
 
         playerDto.nombre = reader.readString();
-        playerDto.playerID = reader.readUint8();
+        playerDto.playerID = reader.readUint32();
         playerDto.raza = reader.readString();
         playerDto.clase = reader.readString();
         playerDto.headId = static_cast<int>(reader.readUint32());
